requirements.txt
pydantic-settings>=2.8.1
mysql-connector-python>=9.2.0
